\name{print.QRank.heter}
\alias{print.QRank.heter}

\title{
Print a QRank.heter object
}
\description{
Print the object of heter.QRank
}
\usage{
\method{print}{QRank.heter}(x, ...)
}

\arguments{
  \item{x}{
Object returned from heter.QRank.
}
  \item{...}{
Optional arguments
}
}

\seealso{
\code{\link{heter.QRank}}
}
